Split packet event out
